Fix TextInputTests
This is suggested fix for https://youtrack.jetbrains.com/issue/CMP-7978
