+++ compose/ui/ui/src/webTest/kotlin/androidx/compose/ui/input/TextInputTests.kt
@@ -16,13 +16,11 @@
 
 package androidx.compose.ui.input
 
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.text.BasicTextField
-import androidx.compose.foundation.text.input.TextFieldState
 import androidx.compose.material.TextField
-import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.SideEffect
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.OnCanvasTests
@@ -30,12 +28,10 @@ import androidx.compose.ui.events.InputEvent
 import androidx.compose.ui.events.InputEventInit
 import androidx.compose.ui.events.keyEvent
 import androidx.compose.ui.focus.FocusRequester
+import androidx.compose.ui.focus.FocusRequester.Companion.FocusRequesterFactory.component1
+import androidx.compose.ui.focus.FocusRequester.Companion.FocusRequesterFactory.component2
 import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.layout.onGloballyPositioned
-import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.sendFromScope
-import androidx.compose.ui.text.TextRange
-import androidx.compose.ui.unit.Density
 import kotlin.test.Ignore
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -53,48 +49,53 @@ import org.w3c.dom.HTMLTextAreaElement
 import org.w3c.dom.events.CompositionEvent
 import org.w3c.dom.events.CompositionEventInit
 import org.w3c.dom.events.Event
-import org.w3c.dom.events.MouseEvent
-import org.w3c.dom.events.MouseEventInit
 
-private class InputChannel(
-    private val channel: Channel<String> = Channel<String>(
+private class InputInteractor(
+    val composeChannel: Channel<String> = Channel<String>(
         1, onBufferOverflow = BufferOverflow.DROP_OLDEST
     )
 ) {
-    lateinit  var htmlInput: HTMLTextAreaElement
-    suspend fun receive() = channel.receive()
+    lateinit var htmlInput: HTMLTextAreaElement
+    suspend fun receive() = composeChannel.receive()
 }
 
-class TextInputTests : OnCanvasTests  {
+class TextInputTests : OnCanvasTests {
+    // delay in web tests called directly will be completely ignored
+    private suspend fun waitFor(millis: Long) {
+        withContext(Dispatchers.Default) { delay(millis) }
+    }
 
-    private fun InputChannel.sendToHtmlInput(vararg events: Event) {
+    private fun InputInteractor.sendToHtmlInput(vararg events: Event) {
         dispatchEvents(htmlInput, *events)
     }
 
-    private fun createTextFieldWithChannel(): InputChannel {
-        val textInputChannel = Channel<String>(
-            1, onBufferOverflow = BufferOverflow.DROP_OLDEST
-        )
-        val inputChannel = InputChannel(textInputChannel)
+    private fun createTextFieldWithChannel(): InputInteractor {
+        val inputInteractor = InputInteractor()
 
         val (firstFocusRequester) = FocusRequester.createRefs()
 
         createComposeWindow {
+            val textState = remember {
+                mutableStateOf("")
+            }
+
             BasicTextField(
-                value = "",
+                value = textState.value,
                 onValueChange = { value ->
-                    textInputChannel.sendFromScope(value)
+                    textState.value = value
+                    inputInteractor.composeChannel.sendFromScope(value)
                 },
                 modifier = Modifier.focusRequester(firstFocusRequester)
             )
 
             LaunchedEffect(Unit) {
                 firstFocusRequester.requestFocus()
-                inputChannel.htmlInput = document.querySelector("textarea") as HTMLTextAreaElement
+                inputInteractor.htmlInput =
+                    document.querySelector("textarea") as HTMLTextAreaElement
             }
         }
 
-        return inputChannel
+        return inputInteractor
     }
 
     @Test
@@ -118,7 +119,11 @@ class TextInputTests : OnCanvasTests  {
             keyEvent("X"),
         )
 
-        assertEquals("stepX", textInputChannel.receive(), "Backspace should delete last symbol typed")
+        assertEquals(
+            "stepX",
+            textInputChannel.receive(),
+            "Backspace should delete last symbol typed"
+        )
     }
 
     @Test
@@ -133,19 +138,21 @@ class TextInputTests : OnCanvasTests  {
             keyEvent("a"),
             compositionStart(),
             beforeInput("insertCompositionText", "a"),
-            keyEvent("a", type = "keyup"),
+            keyEvent("a", type = "keyup", isComposing = true),
             keyEvent("1", code = "Digit1", isComposing = true),
             beforeInput("insertCompositionText", "啊"),
             compositionEnd("啊"),
             keyEvent("1", code = "Digit1", type = "keyup"),
         )
 
+        assertEquals("啊", textInputChannel.receive())
+
         textInputChannel.sendToHtmlInput(
-            keyEvent("X"),
-            keyEvent("X", type = "keyup")
+            keyEvent("x"),
+            keyEvent("x", type = "keyup")
         )
 
-        assertEquals("啊X", textInputChannel.receive())
+        assertEquals("啊x", textInputChannel.receive())
     }
 
     @Ignore // TODO: https://youtrack.jetbrains.com/issue/CMP-7978
@@ -156,7 +163,8 @@ class TextInputTests : OnCanvasTests  {
 
         val keyEvent = keyEvent("1", code = "Digit1")
 
-        delay(50)
+        // We can not change timestamp for js events, so we just add some delay to enforce it
+        waitFor(50)
 
         textInputChannel.sendToHtmlInput(
             compositionStart(),
@@ -166,10 +174,13 @@ class TextInputTests : OnCanvasTests  {
             beforeInput("insertFromComposition", "啊"),
             compositionEnd("啊"),
             keyEvent,
-            keyEvent("1", type="keyup", code = "Digit1"),
+            keyEvent("1", type = "keyup", code = "Digit1"),
         )
 
-        delay(50)
+        assertEquals("啊", textInputChannel.receive())
+
+        // We can not change timestamp for js events, so we just add some delay to enforce it
+        waitFor(100)
 
         textInputChannel.sendToHtmlInput(
             keyEvent("b"),
@@ -223,7 +234,11 @@ class TextInputTests : OnCanvasTests  {
         )
 
         // TODO: this does not behave as desktop, ideally we should have "abc" here
-        assertEquals("bc", textInputChannel.receive(), "Repeat mode should be resolved as Accent Dialogue")
+        assertEquals(
+            "bc",
+            textInputChannel.receive(),
+            "Repeat mode should be resolved as Accent Dialogue"
+        )
 
         textInputChannel.sendToHtmlInput(
             keyEvent("a"),
@@ -260,7 +275,10 @@ class TextInputTests : OnCanvasTests  {
             keyEvent("c")
         )
 
-        assertTrue(Regex("a+bc").matches(textInputChannel.receive()), "Repeat mode should be resolved as Default")
+        assertTrue(
+            Regex("a+bc").matches(textInputChannel.receive()),
+            "Repeat mode should be resolved as Default"
+        )
     }
 
     @Test
@@ -327,7 +345,6 @@ class TextInputTests : OnCanvasTests  {
     }
 
 
-
     @Test
     fun keyboardEventPassedToTextField() = runTest {
 
@@ -389,90 +406,17 @@ class TextInputTests : OnCanvasTests  {
 
         assertEquals("step2", textInputChannel.receive())
     }
-
-    @Test
-    fun canSelectUsingMouse() = runTest {
-        val syncChannel = Channel<TextRange?>(
-            1, onBufferOverflow = BufferOverflow.DROP_OLDEST
-        )
-
-        var textFieldWidth = 0
-
-        createComposeWindow {
-            val textState = remember { TextFieldState("qwerty 1234567") }
-            val focusRequester = remember { FocusRequester() }
-
-            CompositionLocalProvider(LocalDensity provides Density(2f)) {
-                Column {
-                    TextField(state = textState, modifier = Modifier.focusRequester(focusRequester).onGloballyPositioned {
-                        textFieldWidth = it.size.width
-                    })
-
-                    LaunchedEffect(textState.selection) {
-                        focusRequester.requestFocus()
-                        syncChannel.send(textState.selection)
-                    }
-                }
-            }
-
-            SideEffect {
-                focusRequester.requestFocus()
-            }
-        }
-
-        yield()
-
-        var selection = syncChannel.receive()
-        assertEquals(TextRange(14, 14), selection)
-        assertTrue(textFieldWidth > 0, "TextField width should be positive")
-
-        val canvas = getCanvas()
-        canvas.dispatchEvent(MouseEvent("mouseenter"))
-        yield()
-        canvas.dispatchEvent(MouseEvent("mousedown", MouseEventInit(clientX = 8, clientY = 20, buttons = 1, button = 1)))
-        yield()
-        canvas.dispatchEvent(MouseEvent("mouseup", MouseEventInit(clientX = 8, clientY = 20, buttons = 0, button = 1)))
-
-        selection = syncChannel.receive()
-        assertEquals(TextRange(0, 0), selection)
-
-        val textArea = document.querySelector("textarea")
-        assertIs<HTMLTextAreaElement>(textArea)
-
-        val textAreaRect = textArea.getBoundingClientRect()
-        // Do a manual hit-test
-        val elementsAtPos = document.elementsFromPoint(
-            textAreaRect.left + textAreaRect.width / 2 ,
-            textAreaRect.top + textAreaRect.height / 2
-        )
-
-        // We expect the canvas to be on the top despite the coordinates match the textarea.
-        // So it will be the first to process all the point inputs
-        assertEquals(canvas, elementsAtPos[0], "First element under mouse supposed to be canvas")
-        assertTrue(elementsAtPos.toList().any { it == textArea }) // such a weird check to make the test common for js and wasm
-        withContext(Dispatchers.Default) {
-            delay(250) // to separate the mouse events
-        }
-
-
-        // Try to select the text using mouse:
-        val startX = textAreaRect.left.toInt() + 1
-        val startY = textAreaRect.top.toInt() + 8
-        val endX = startX + textFieldWidth
-        canvas.dispatchEvent(MouseEvent("mousemove", MouseEventInit(clientX = startX, clientY = startY, buttons = 1, button = 1)))
-        canvas.dispatchEvent(MouseEvent("mousedown", MouseEventInit(clientX = startX, clientY = startY, buttons = 1, button = 1)))
-        canvas.dispatchEvent(MouseEvent("mousemove", MouseEventInit(clientX = endX, clientY = startY, buttons = 1, button = 1)))
-        canvas.dispatchEvent(MouseEvent("mouseup", MouseEventInit(clientX = endX, clientY = startY, buttons = 0, button = 1)))
-
-        selection = syncChannel.receive()
-        assertEquals(TextRange(0, 14), selection)
-    }
 }
 
 
-private fun compositionStart(data: String = "")  = CompositionEvent("compositionstart", CompositionEventInit(data = data))
-private fun compositionEnd(data: String) = CompositionEvent("compositionend", CompositionEventInit(data = data))
-private fun beforeInput(inputType: String, data: String?) = InputEvent("beforeinput", InputEventInit(inputType = inputType, data = data))
+private fun compositionStart(data: String = "") =
+    CompositionEvent("compositionstart", CompositionEventInit(data = data))
+
+private fun compositionEnd(data: String) =
+    CompositionEvent("compositionend", CompositionEventInit(data = data))
+
+private fun beforeInput(inputType: String, data: String?) =
+    InputEvent("beforeinput", InputEventInit(inputType = inputType, data = data))
 
 private fun mobileKeyDown() = keyEvent(type = "keydown", key = "Unidentified", code = "")
 private fun mobileKeyUp() = keyEvent(type = "keydown", key = "Unidentified", code = "")
